requirements.txt
django==2.2.10
pymysql
mysqlclient
djangorestframework
django-redis
django-cors-headers











